Added Collab and Fixed Scoring System

+++ Assets/ScoreText.cs
@@ -8,6 +8,6 @@ public class ScoreText : MonoBehaviour {
 	
 	// Update is called once per frame
 	void Update () {
-		GetComponent<Text>().text = PlayerPrefs.GetInt(namestring)+"";
+		GetComponent<Text>().text = PlayerPrefs.GetFloat(namestring)+"";
 	}
 }
